+++ CHANGES.rst
@@ -1,6 +1,24 @@
 Changes
 =======
 
+0.2.0 (unreleased)
+------------------
+
+- Add ``Visibility``, ``Collapsible`` and ``Button`` widgets.
+  [rnix]
+
+- Add ``query_elem``, ``get_elem`` and ``set_visible`` utility functions.
+  [rnix]
+
+- Add ``form`` module.
+  [rnix]
+
+- Add ``listener`` module.
+  [rnix]
+
+- Add ``websocket`` module.
+  [rnix]
+
 0.1.0 (2021-12-13)
 ------------------